Reorganize Authentication API tests; no changes to actual tests or coverage
